Output type changed to float in the test reference file of Shape Operator

+++ tmva/sofie/test/input_models/references/Shape.ref.hxx
@@ -1,5 +1,5 @@
 namespace Shape_ExpectedOutput{
-	float outputs[] = {
+	int outputs[] = {
         1, 2
 	};
 } // namespace Shape_ExpectedOutput
